requirements.txt
pandas
numpy
matplotlib
mysql-connector-python
sqlalchemy
streamlit
